+++ ui/frontend-lib/src/permissions/components/policies/PolicyApiCreateDialog.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from "react";
+import { useCallback } from "react";
 
 import { Controller, useForm } from "react-hook-form";
