+++ ui/frontend-lib/src/tasks/pages/Tasks.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useMemo } from "react";
 
 import { useNavigate } from "react-router";
